+++ changelog.md
@@ -3,3 +3,6 @@
 ## Additions
 - Add diesel
 - Add diesel diesel generator
+
+## Changes
+- Rover now need diesel instead of fuel